requirements-test.txt
pytest==2.5.1
pytest-pep8==1.0.5
pytest-django=2.4
coverage==3.7.1
python-coveralls==2.4.2
